third_party/blink/web_tests/fast/events/keyboardevent-location-constants-expected.txt

Tests that the DOM_KEY_LOCATION_* constants are defined on KeyboardEvent

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS KeyboardEvent.DOM_KEY_LOCATION_STANDARD is 0x00
PASS KeyboardEvent.DOM_KEY_LOCATION_LEFT is 0x01
PASS KeyboardEvent.DOM_KEY_LOCATION_RIGHT is 0x02
PASS KeyboardEvent.DOM_KEY_LOCATION_NUMPAD is 0x03
PASS successfullyParsed is true

TEST COMPLETE
